requirements.txt
azure-ai-projects>=1.0.0
azure-ai-agents>=1.0.0
azure-identity>=1.14.0
python-dotenv>=1.0.0
asyncio
